[clang][NFC] Move more things out of SemaChecking.cpp
